requirements.txt
requests
python-dotenv
typer
rich

jsonschema
